Refactor: Separate into components, hooks, and utils
